[WASI] Rename the file name.
