+++ include/host/wasi_nn/wasinnfunc.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "common/errcode.h"
-#include "host/wasi_nn/wasinncontext.h"
+#include "host/wasi_nn/wasinnenv.h"
 #include "runtime/hostfunc.h"
 #include "runtime/instance/memory.h"
